docker management scripts fix

+++ deployment/indexers-testnet/stop.sh
@@ -2,4 +2,8 @@
 
 chain="$1"
 
-docker-compose -f docker-compose-indexer-$chain.yaml -p indexer-$chain down
+if [[ -z $2 ]] ; then
+   docker-compose -f docker-compose-indexer-$chain.yaml -p indexer-$chain down   
+else
+   docker-compose -f docker-compose-indexer-$chain.yaml -p indexer-$chain stop "${@:2}"
+fi
